使用C实现A*算法，提升计算速度

+++ app/src/main/java/com/example/autopieces/logic/combat/search/Node.kt
@@ -3,11 +3,11 @@ package com.example.autopieces.logic.combat.search
 class Node(
     var x : Int = 0,
     var y : Int = 0,
-    var f : Float = 0f,
+    var f : Int = 0,
     //从起点A移动到指定位置T的移动代价
-    var g : Float = 0f,
+    var g : Int = 0,
     //从指定位置T移动到终点的估算成本
-    var h : Float = 0f,
+    var h : Int = 0,
     var parent : Node? = null
 ) {
     override fun equals(other: Any?): Boolean {
